apply changes from reviewer
