+++ tests/test_codegen_file_object.py
@@ -17,7 +17,9 @@ from arx import codegen as codegen_module
 from arx.codegen import ArxBuilder, ArxVisitor
 from arx.io import ArxIO
 from arx.lexer import Lexer
+from arx.main import FileImportResolver, get_module_name_from_file_path
 from arx.parser import Parser
+from irx.analysis.module_interfaces import ParsedModule
 from irx.builder.runtime.registry import RuntimeFeatureState
 from llvmlite import binding as llvm
 
@@ -133,6 +135,77 @@ def test_object_generation(code: str) -> None:
     bin_path = TMP_PATH / "testtmp"
     ir.build(module_ast, str(bin_path))
     bin_path.unlink()
+
+
+@pytest.mark.skipif(not HAS_CLANG, reason="clang is required for object build")
+def test_namespace_import_program_builds_and_runs(tmp_path: Path) -> None:
+    """
+    title: Namespace-imported module calls should survive full build and run.
+    parameters:
+      tmp_path:
+        type: Path
+    """
+    project_root = tmp_path / "workspace"
+    package_dir = project_root / "src" / "samplepkg"
+    package_dir.mkdir(parents=True)
+
+    (project_root / ".arxproject.toml").write_text(
+        '[project]\nname = "samplepkg"\nversion = "0.1.0"\n'
+        '[environment]\nkind = "conda"\nname = "samplepkg"\n'
+        '[build]\nsrc_dir = "src"\nentry = "samplepkg/__init__.x"'
+        '\nout_dir = "build"\n',
+        encoding="utf-8",
+    )
+    (package_dir / "__init__.x").write_text(
+        dedent(
+            """
+            import samplepkg.stats as stats
+
+            fn main() -> i32:
+              return stats.sum2(1, 2)
+            """
+        ).lstrip(),
+        encoding="utf-8",
+    )
+    (package_dir / "stats.x").write_text(
+        dedent(
+            """
+            fn sum2(a: i32, b: i32) -> i32:
+              return a + b
+            """
+        ).lstrip(),
+        encoding="utf-8",
+    )
+
+    root_file = package_dir / "__init__.x"
+    ArxIO.file_to_buffer(str(root_file))
+    module_ast = Parser().parse(
+        Lexer().lex(),
+        get_module_name_from_file_path(str(root_file)),
+    )
+    assert isinstance(module_ast, astx.Module)
+
+    root = ParsedModule(
+        key=module_ast.name,
+        ast=module_ast,
+        display_name=module_ast.name,
+        origin=str(root_file),
+    )
+    resolver = FileImportResolver((str(root_file),))
+
+    bin_path = tmp_path / "namespace_import_program"
+    ArxBuilder().build_modules(root, resolver, str(bin_path))
+
+    result = subprocess.run(
+        [str(bin_path)],
+        check=False,
+        capture_output=True,
+        text=True,
+    )
+
+    assert result.returncode == 3
+    assert result.stdout == ""
+    assert result.stderr == ""
 
 
 @pytest.mark.skipif(not HAS_CLANG, reason="clang is required for object build")
